Added support for gravity for image rendering
this commit adds support to allow the user to specify the gravity for
the background image. Gravity have similar effect to ImageMagick's
gravity parameter for cropping. In short, it helps the user specify the
region of the image to display in the background.

Previously, swaybg defaults to the center of the image after the scaling
done by cairo which might not be ideal.